requirements.txt
flake8==5.0.4
fastapi==0.86.0
kafka-python==2.0.2
pre-commit==2.20.0
psycopg2-binary==2.9.3
pyflakes==2.5.0
pytest==7.1.0
pylint==2.15.3
python-dotenv==0.20.0
SQLAlchemy==1.4.37
supertokens-python==0.11.5
black==22.8.0